third_party/WebKit/LayoutTests/external/wpt/html/browsers/browsing-the-web/unloading-documents/001-expected.txt

CONSOLE ERROR: line 21: Uncaught TypeError: do_test is not a function
This is a testharness.js-based test.
FAIL document.open in unload assert_equals: expected "0123456789" but got "012389"
Harness: the test ran to completion.
